// client-id/src/app/core/models/inspection-popup/ipt-elecmeter-model.ts
export class IptElecMeterModel{
    athrzSn: number;
    athrzDetailSn: number;
    fomCode: string;
    fomCodeName: string;
    estbSeCode: string;
    estbSeCodeName: string;
    upptNdLnFrmlaCode: string;
    upptNdLnFrmlaCodeName: string;
    stdrVltgeCode: string;
    stdrVltgeCodeName: string;
    stdrErcrtCode: string;
    stdrErcrtCodeName: string;
    mxmmPermErcrtVal: number;
    validGradCode: string;
    validGradCodeName: string;
    unavblGradCode: string;
    unavblGradCodeName: string;
    mesurDrcCode: string;
    mesurDrcCodeName: string;
    fomNm: string;
    conectStrctCode: string;
    conectStrctCodeName: string;
    chgEstbSeCode: string;
    chgEstbSeCodeName: string;
    mode: string;
    printErcrtVal: string;
    swVersion: string;
    constructor(){
        this.athrzSn = 0;
        this.athrzDetailSn = 0;
        this.fomCode = null;
        this.fomCodeName = null;
        this.estbSeCode = null;
        this.estbSeCodeName = null;
        this.upptNdLnFrmlaCode = null;
        this.upptNdLnFrmlaCodeName = null;
        this.stdrVltgeCode = null;
        this.stdrVltgeCodeName = null;
        this.stdrErcrtCode = null;
        this.stdrErcrtCodeName = null;
        this.mxmmPermErcrtVal = 0;
        this.validGradCode = null;
        this.validGradCodeName = null;
        this.unavblGradCode = null;
        this.unavblGradCodeName = null;
        this.mesurDrcCode = null;
        this.mesurDrcCodeName = null;
        this.fomNm = null;
        this.conectStrctCode = null;
        this.conectStrctCodeName = null;
        this.chgEstbSeCode = null;
        this.chgEstbSeCodeName = null;
        this.mode = null;
        this.printErcrtVal = null;
        this.swVersion = null;
    }
}